update: 第四期：判题服务与代码沙箱模块

+++ src/main/java/com/yuhao/deepoj/model/dto/problem/JudgeConfig.java
@@ -11,20 +11,20 @@ public class JudgeConfig {
     /**
      * 时间限制(ms)
      */
-    private Long timeLimit;
+    private Integer timeLimit;
 
     /**
      * 内存限制(MB)
      */
-    private Long memoryLimit;
+    private Integer memoryLimit;
 
     /**
      * 堆栈限制(MB)
      */
-    private Long stackLimit;
+    private Integer stackLimit;
 
     /**
      * 代码长度限制(B)
      */
-    private Long codeLengthLimit;
+    private Integer codeLengthLimit;
 }
